+++ src/test/java/itest/agent/AgentTargetAnalysisIT.java
@@ -13,80 +13,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package itest;
+package itest.agent;
 
 import static io.restassured.RestAssured.given;
 
 import java.io.IOException;
-import java.net.URI;
-import java.time.Duration;
-import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Executors;
-import java.util.concurrent.LinkedBlockingDeque;
 import java.util.concurrent.TimeUnit;
-import java.util.function.Predicate;
 
-import io.cryostat.resources.AgentApplicationResource;
-
-import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.QuarkusIntegrationTest;
 import io.restassured.http.ContentType;
 import io.smallrye.mutiny.TimeoutException;
 import io.vertx.core.MultiMap;
 import io.vertx.core.json.JsonObject;
-import itest.bases.HttpClientTest;
-import itest.resources.S3StorageResource;
-import jakarta.websocket.ClientEndpoint;
-import jakarta.websocket.ContainerProvider;
 import jakarta.websocket.DeploymentException;
-import jakarta.websocket.OnMessage;
-import jakarta.websocket.Session;
 import org.hamcrest.Matchers;
-import org.jboss.logging.Logger;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledIf;
 
 @QuarkusIntegrationTest
-@QuarkusTestResource(value = AgentApplicationResource.class, restrictToAnnotatedClass = true)
-@QuarkusTestResource(value = S3StorageResource.class, restrictToAnnotatedClass = true)
 @EnabledIf("enabled")
-public class TargetAnalysisIT extends HttpClientTest {
-
-    public static final String TEMPLATE_CONTINUOUS = "template=Continuous,type=TARGET";
-
-    static WebSocketClient WS_CLIENT;
-    static Session WS_SESSION;
-
-    public static boolean enabled() {
-        String arch = Optional.ofNullable(System.getenv("CI_ARCH")).orElse("").trim();
-        boolean ci = Boolean.valueOf(System.getenv("CI"));
-        return !ci || (ci && "amd64".equalsIgnoreCase(arch));
-    }
-
-    @BeforeAll
-    static void setupWebSocketClient() throws IOException, DeploymentException {
-        WS_CLIENT = new WebSocketClient();
-        WS_SESSION =
-                ContainerProvider.getWebSocketContainer()
-                        .connectToServer(
-                                WS_CLIENT, URI.create("ws://localhost:8081/api/notifications"));
-    }
-
-    @BeforeEach
-    void clearWebSocketNotifications() {
-        WS_CLIENT.msgQ.clear();
-    }
-
-    @AfterAll
-    static void tearDownWebSocketClient() throws IOException {
-        WS_SESSION.close();
-    }
+public class AgentTargetAnalysisIT extends AgentTestBase {
 
     @Test
     void testGetAgentTargetReport()
@@ -96,13 +45,13 @@ public class TargetAnalysisIT extends HttpClientTest {
                     TimeoutException,
                     ExecutionException,
                     java.util.concurrent.TimeoutException {
-        Target agent = getAgentTarget();
-        long targetId = agent.id;
+        Target agent = waitForDiscovery();
+        long targetId = target.id();
         String archivedRecordingName = null;
         long recordingId = -1;
         try {
             recordingId =
-                    startRecording(targetId, "targetAnalysisReportRecording", TEMPLATE_CONTINUOUS);
+                    startRecording(targetId, "targetAnalysisReportRecording", CONTINUOUS_TEMPLATE);
 
             Thread.sleep(5_000);
 
@@ -122,9 +71,7 @@ public class TargetAnalysisIT extends HttpClientTest {
                                                 .and()
                                                 .assertThat()
                                                 // 202 Indicates report generation is in progress
-                                                // and sends an
-                                                // intermediate
-                                                // response.
+                                                // and sends an intermediate response.
                                                 .statusCode(202)
                                                 .contentType(ContentType.TEXT)
                                                 .body(Matchers.any(String.class))
@@ -155,14 +102,14 @@ public class TargetAnalysisIT extends HttpClientTest {
                     "ReportSuccess",
                     o ->
                             Objects.equals(
-                                    agent.jvmId, o.getJsonObject("message").getString("jvmId")));
+                                    agent.jvmId(), o.getJsonObject("message").getString("jvmId")));
         } finally {
             if (archivedRecordingName != null) {
                 given().log()
                         .all()
                         .when()
                         .pathParams(
-                                "connectUrl", agent.connectUrl, "filename", archivedRecordingName)
+                                "connectUrl", agent.connectUrl(), "filename", archivedRecordingName)
                         .delete("/api/beta/recordings/{connectUrl}/{filename}")
                         .then()
                         .log()
@@ -212,100 +159,5 @@ public class TargetAnalysisIT extends HttpClientTest {
                 .extract()
                 .jsonPath()
                 .getLong("remoteId");
-    }
-
-    private Target getAgentTarget() throws InterruptedException {
-        final long timeout = 60_000;
-        final long start = System.currentTimeMillis();
-        long now = start;
-        do {
-            now = System.currentTimeMillis();
-            var targets =
-                    given().log()
-                            .all()
-                            .when()
-                            .get("/api/v4/targets")
-                            .then()
-                            .log()
-                            .all()
-                            .and()
-                            .assertThat()
-                            .statusCode(200)
-                            .and()
-                            .extract()
-                            .body()
-                            .as(Target[].class);
-            for (var target : targets) {
-                if (Objects.equals(target.alias, AgentApplicationResource.ALIAS)) {
-                    return target;
-                }
-            }
-            Thread.sleep(5_000);
-        } while (now < start + timeout);
-        throw new IllegalStateException();
-    }
-
-    private record Target(
-            long id,
-            String jvmId,
-            String connectUrl,
-            String alias,
-            Annotations annotations,
-            List<KeyValue> labels,
-            boolean agent) {}
-
-    private record Annotations(List<KeyValue> platform, List<KeyValue> cryostat) {}
-
-    private record KeyValue(String key, String value) {}
-
-    protected JsonObject expectWebSocketNotification(String category)
-            throws IOException, DeploymentException, InterruptedException, TimeoutException {
-        return expectWebSocketNotification(category, Duration.ofSeconds(30), v -> true);
-    }
-
-    protected JsonObject expectWebSocketNotification(String category, Duration timeout)
-            throws IOException, DeploymentException, InterruptedException, TimeoutException {
-        return expectWebSocketNotification(category, timeout, v -> true);
-    }
-
-    protected JsonObject expectWebSocketNotification(
-            String category, Predicate<JsonObject> predicate)
-            throws IOException, DeploymentException, InterruptedException, TimeoutException {
-        return expectWebSocketNotification(category, Duration.ofSeconds(30), predicate);
-    }
-
-    protected JsonObject expectWebSocketNotification(
-            String category, Duration timeout, Predicate<JsonObject> predicate)
-            throws IOException, DeploymentException, InterruptedException, TimeoutException {
-        long now = System.nanoTime();
-        long deadline = now + timeout.toNanos();
-        logger.infov(
-                "waiting up to {0} for a WebSocket notification with category={1}",
-                timeout, category);
-        do {
-            now = System.nanoTime();
-            String msg = WS_CLIENT.msgQ.poll(1, TimeUnit.SECONDS);
-            if (msg == null) {
-                continue;
-            }
-            JsonObject obj = new JsonObject(msg);
-            String msgCategory = obj.getJsonObject("meta").getString("category");
-            if (category.equals(msgCategory) && predicate.test(obj)) {
-                return obj;
-            }
-        } while (now < deadline);
-        throw new TimeoutException();
-    }
-
-    @ClientEndpoint
-    private static class WebSocketClient {
-        private final LinkedBlockingDeque<String> msgQ = new LinkedBlockingDeque<>();
-        private final Logger logger = Logger.getLogger(getClass());
-
-        @OnMessage
-        void message(String msg) {
-            logger.info(msg);
-            msgQ.add(msg);
-        }
     }
 }
